# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feloxi API",
    "description": "Celery task monitoring and management API",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Org selection needed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgPickerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Return the current user info (used by frontend to check if session is active).",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/v1/auth/switch-org": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "switch_org",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not a member"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "description": "Successful auth response. Tokens are set as HttpOnly cookies, NOT in the body.",
        "required": [
          "user",
          "expires_in"
        ],
        "properties": {
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "tenant_slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional: if the user belongs to multiple orgs, they must specify which one."
          }
        }
      },
...
      "OrgPickerResponse": {
        "type": "object",
        "description": "Returned when the user's email is in multiple orgs and no slug was provided.",
        "required": [
          "needs_org_selection",
          "organizations"
        ],
        "properties": {
          "needs_org_selection": {
            "type": "boolean"
          },
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgSummary"
            }
          }
        }
      },
      "OrgSummary": {
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "tenant_name",
          "tenant_slug",
          "email",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "tenant_name": {
            "type": "string"
          },
          "tenant_slug": {
            "type": "string"
          }
        }
      },
...
      "SwitchOrgRequest": {
        "type": "object",
        "description": "Switch to a different org (re-issues tokens for that org).",
        "required": [
          "tenant_slug"
        ],
        "properties": {
          "tenant_slug": {
            "type": "string"
          }
        }
      },
...
      "UserInfo": {
        "type": "object",
        "required": [
          "id",
          "email",
          "tenant_id",
          "tenant_slug",
          "roles"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_slug": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}